+++ superwall/src/main/java/com/superwall/sdk/models/attribution/AttributionProvider.kt
@@ -1,7 +1,5 @@
 package com.superwall.sdk.models.attribution
 
-import com.google.android.gms.ads.identifier.AdvertisingIdClient
-import com.superwall.sdk.Superwall
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -106,6 +104,5 @@ enum class AttributionProvider(
     GOOGLE_ADS("google_ads"),
 
     @SerialName("custom")
-    CUSTOM("custom");
-
+    CUSTOM("custom"),
 }